clion linting

+++ lib/MD5.c
@@ -338,7 +338,7 @@ void md5_append(md5_state_t *pms, const md5_byte_t *data, unsigned nbytes) {
 	return;
 
     /* Update the message length. */
-    pms->count[1] += (unsigned)nbytes >> 29;
+    pms->count[1] += nbytes >> 29;
     pms->count[0] += nbits;
     if (pms->count[0] < nbits)
 	pms->count[1]++;
